pyproject.toml
[tool.poetry]
name = "python-fastapi-entry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.23"
aiomysql = "^0.2.0"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21.1"
aiosqlite = "^0.19.0"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
